pyproject.toml
[project]
name = "viz-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.18.0",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "pdfplumber>=0.11.0",
    "requests>=2.31.0",
    "openpyxl>=3.1.0",
    "pytest",
]
